feat(TS): move info-comp types to package folder
