+++ index.d.ts
@@ -1,51 +1,3 @@
-declare module 'blockchain-info-components' {
-  import { StatelessComponent, CSSProperties } from 'react'
-  import { DefaultTheme } from 'styled-components'
-
-  // type-coverage:ignore-next-line
-  export function BlockchainLoader(...args: any): any
-  // type-coverage:ignore-next-line
-  export function Button(...args: any): any
-  // type-coverage:ignore-next-line
-  export function FontGlobalStyles(...args: any): any
-  // type-coverage:ignore-next-line
-  export function Icon(...args: any): any
-  // type-coverage:ignore-next-line
-  export function IconGlobalStyles(...args: any): any
-  // type-coverage:ignore-next-line
-  export function Image(...args: any): any
-  // type-coverage:ignore-next-line
-  export function Link(...args: any): any
-  // type-coverage:ignore-next-line
-  export function Modal(...args: any): any
-  // type-coverage:ignore-next-line
-  export function ModalBody(...args: any): any
-  // type-coverage:ignore-next-line
-  export function ModalHeader(...args: any): any
-  // type-coverage:ignore-next-line
-  export function Palette(theme: string): DefaultTheme
-  // type-coverage:ignore-next-line
-  export function SkeletonRectangle(...args: any): any
-  // type-coverage:ignore-next-line
-  export function Table(...args): any
-  // type-coverage:ignore-next-line
-  export function TableCell(...args): any
-  // type-coverage:ignore-next-line
-  export function TableHeader(...args): any
-  // type-coverage:ignore-next-line
-  export function TableRow(...args): any
-  // type-coverage:ignore-next-line
-  export const Text: StatelessComponent<{
-    color?: keyof DefaultTheme
-    size?: string
-    weight?: number
-    style?: CSSProperties
-    uppercase?: boolean
-    lineHeight?: string
-  }>
-  // type-coverage:ignore-next-line
-  export function TextGroup(...args: any): any
-}
 declare module 'daggy' {
   import {
     RemoteFailure,
